+++ test/posthog_flutter_platform_interface_fake.dart
@@ -2,6 +2,7 @@ import 'package:posthog_flutter/src/posthog_flutter_platform_interface.dart';
 
 class PosthogFlutterPlatformFake extends PosthogFlutterPlatformInterface {
   String? screenName;
+  OnFeatureFlagsCallback? registeredOnFeatureFlagsCallback;
 
   @override
   Future<void> screen({
@@ -9,5 +10,10 @@ class PosthogFlutterPlatformFake extends PosthogFlutterPlatformInterface {
     Map<String, Object>? properties,
   }) async {
     this.screenName = screenName;
+  }
+
+  @override
+  void onFeatureFlags(OnFeatureFlagsCallback callback) {
+    registeredOnFeatureFlagsCallback = callback;
   }
 }